requirements.txt
requests
pymysql
sshtunnel
cryptography
pandas
streamlit
plotly
scipy
scikit-learn
numpy
python-dotenv
